requirements.txt
# Core ML/NLP Libraries
faiss-cpu==1.8.0
sentence-transformers==2.7.0
#langchain==0.1.16
#langchain-community==0.0.34
#langchain-openai==0.1.3
transformers==4.40.1
torch==2.2.1

# Data Handling & PDF Processing
pypdf==4.2.0
numpy==1.26.4
PyYAML==6.0.1

# Force Pydantic v1 for langchain_openai compatibility
pydantic==1.10.12

# Development & Linting
ruff==0.4.2
pytest==8.1.1
rank-bm25==0.2.2


# PDF rendering → PIL images
pdf2image==1.16.3
pillow==9.5.0

# OCR engine
pytesseract==0.3.10

